+++ react/DrawerAnimation.js
@@ -1,3 +1,0 @@
-import DrawerAnimation from './components/DrawerAnimation/index'
-
-export default DrawerAnimation